refactored account, endpoints, and implemented contact list endpoint
